requirements.txt
# Python >= 3.11 required
streamlit>=1.30.0
openpyxl>=3.1.0
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# dev / test
pytest>=7.4.0
